moves lance tests
