+++ tests/load/lance/conftest.py
@@ -1,0 +1,1 @@
+from tests.load.lance.lance_utils import lance_rest_server, cleanup_lance_namespace_root
